+++ CHANGES.md
@@ -1,8 +1,8 @@
 **Added:**
-* option to change install directory in InstallerConfigWindow
+* 
 
 **Changed:**
-* progress on the no-setting installer. untested
+* made some progress on installer again
 
 **Fixed:**
 * none